requirements.txt
Flask
gunicorn
torch
numpy
opencv-python
Pillow
